requirements.txt
python-docx==0.8.11
PyPDF2==3.0.1
markdown==3.4.3
Flask==2.3.3
pypandoc==1.11 